refactor: replace flat tests.rs with tests/ subdirectory structure for arabic, clean, dollar, fixed, roman
Migrates each function's tests from a single flat tests.rs to the
required tests/mod.rs + tests/success.rs + tests/failure.rs pattern
used by all other M2 functions.
