+++ crates/core/src/eval/functions/text/arabic/tests/mod.rs
@@ -1,0 +1,2 @@
+mod success;
+mod failure;
