+++ crates/core/src/eval/functions/text/arabic/tests/success.rs
@@ -1,5 +1,5 @@
-use super::*;
-use crate::types::{ErrorKind, Value};
+use super::super::*;
+use crate::types::Value;
 
 fn run(s: &str) -> Value {
     arabic_fn(&[Value::Text(s.to_string())])
@@ -43,16 +43,6 @@ fn mmmcmxcix() {
 #[test]
 fn empty_returns_zero() {
     assert_eq!(run(""), Value::Number(0.0));
-}
-
-#[test]
-fn invalid_returns_value_error() {
-    assert_eq!(run("ABC"), Value::Error(ErrorKind::Value));
-}
-
-#[test]
-fn no_args_returns_na() {
-    assert_eq!(arabic_fn(&[]), Value::Error(ErrorKind::NA));
 }
 
 #[test]
