+++ crates/core/src/eval/functions/text/clean/tests/failure.rs
@@ -1,0 +1,7 @@
+use super::super::*;
+use crate::types::{ErrorKind, Value};
+
+#[test]
+fn no_args_returns_na() {
+    assert_eq!(clean_fn(&[]), Value::Error(ErrorKind::NA));
+}
