+++ crates/core/src/eval/functions/text/dollar/tests/mod.rs
@@ -1,0 +1,2 @@
+mod success;
+mod failure;
