+++ crates/core/src/eval/functions/text/fixed/tests/mod.rs
@@ -1,0 +1,2 @@
+mod success;
+mod failure;
